Carrito-service ahora comprueba existencia de id de producto y usuario con sus services respectivos mediante OpenFeing

+++ carrito_service/src/main/java/com/pokemon/carrito_service/Controller/CarritoController.java
@@ -31,7 +31,7 @@ public class CarritoController {
 
     // Carritos por usuario
     @GetMapping("/usuario/{id}")
-    public ResponseEntity<List<Carrito>> listarPorUsuario(@PathVariable Integer id) {
+    public ResponseEntity<List<Carrito>> listarPorUsuario(@PathVariable Long id) {
 
         List<Carrito> carritos = carritoService.listarPorUsuario(id);
