+++ carrito_service/src/main/java/com/pokemon/carrito_service/Model/Carrito.java
@@ -18,7 +18,7 @@ public class Carrito {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer idCarrito;
 
-    private Integer idUsuario;
+    private Long idUsuario;
 
     private LocalDateTime fechaCreacion;
 }
